requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bandit>=1.7.5
bandit-sarif-formatter>=1.0.3
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
flake8>=6.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
locust>=2.35.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
